# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Docseq DMS API
  description: |
    Sistema de Gestão de Documentos Municipais.
    
    **Base URL:** `http://localhost:8080`
  version: 1.0.0
  contact:
    name: Docseq DMS
servers:
  - url: http://localhost:8080/api/v1
    description: Desenvolvimento local
...
paths:
...
  /municipalities:
    post:
      summary: Criar município
      operationId: createMunicipality
      tags: [municipalities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMunicipalityInput'
            examples:
              passagem_pb:
                summary: Passagem - PB
                value:
                  name: "Passagem"
                  uf: "PB"
                  imageUrl: "https://example.com/brasoes/passagem-pb.png"
              joao_pessoa:
                summary: João Pessoa - PB (sem imagem)
                value:
                  name: "João Pessoa"
                  uf: "PB"
      responses:
        '201':
          description: Município criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Passagem"
                  uf: "PB"
                  imageUrl: "https://example.com/brasoes/passagem-pb.png"
                  createdAt: "2025-01-15T10:30:00Z"
                  updatedAt: "2025-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

    # ─────────────────────────────────────────────
    # GET /municipalities
    # ─────────────────────────────────────────────
    get:
      summary: Listar municípios (paginado)
      operationId: listMunicipalities
      tags: [municipalities]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Lista paginada de municípios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
              example:
                success: true
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Passagem"
                    uf: "PB"
                    imageUrl: "https://example.com/brasoes/passagem-pb.png"
                    createdAt: "2025-01-15T10:30:00Z"
                    updatedAt: "2025-01-15T10:30:00Z"
                  - id: "660e8400-e29b-41d4-a716-446655440001"
                    name: "Patos"
                    uf: "PB"
                    createdAt: "2025-01-15T11:00:00Z"
                    updatedAt: "2025-01-15T11:00:00Z"
                pagination:
                  page: 1
                  pageSize: 20
                  total: 2
                  totalPages: 1
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # GET /municipalities/trash
  # ─────────────────────────────────────────────
  /municipalities/trash:
    get:
      summary: Listar municípios removidos (lixeira)
      operationId: listDeletedMunicipalities
      tags: [municipalities]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Lista paginada de municípios removidos por soft-delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
              example:
                success: true
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Passagem"
                    uf: "PB"
                    imageUrl: "https://example.com/brasoes/passagem-pb.png"
                    createdAt: "2025-01-15T10:30:00Z"
                    updatedAt: "2025-01-15T10:30:00Z"
                pagination:
                  page: 1
                  pageSize: 20
                  total: 1
                  totalPages: 1
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # GET /municipalities/:id
  # ─────────────────────────────────────────────
  /municipalities/{id}:
    parameters:
      - $ref: '#/components/parameters/MunicipalityID'

    get:
      summary: Buscar município por ID
      operationId: getMunicipalityById
      tags: [municipalities]
      responses:
        '200':
          description: Município encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Passagem"
                  uf: "PB"
                  imageUrl: "https://example.com/brasoes/passagem-pb.png"
                  createdAt: "2025-01-15T10:30:00Z"
                  updatedAt: "2025-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    # ─────────────────────────────────────────────
    # PATCH /municipalities/:id
    # ─────────────────────────────────────────────
    patch:
      summary: Atualizar município (parcial)
      operationId: updateMunicipality
      tags: [municipalities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMunicipalityInput'
            examples:
              rename:
                summary: Alterar nome
                value:
                  name: "Passagem Nova"
              change_image:
                summary: Alterar imagem
                value:
                  imageUrl: "https://example.com/brasoes/novo.png"
              full_update:
                summary: Atualizar tudo
                value:
                  name: "Passagem"
                  uf: "PB"
                  imageUrl: "https://example.com/brasoes/passagem-pb.png"
      responses:
        '200':
          description: Município atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

    # ─────────────────────────────────────────────
    # DELETE /municipalities/:id
    # ─────────────────────────────────────────────
    delete:
      summary: Remover município (soft-delete)
      operationId: deleteMunicipality
      tags: [municipalities]
      responses:
        '204':
          description: Município removido com sucesso (sem conteúdo)
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # DELETE /municipalities/:id/hard
  # ─────────────────────────────────────────────
  /municipalities/{id}/hard:
    parameters:
      - $ref: '#/components/parameters/MunicipalityID'

    delete:
      summary: Remover município definitivamente (hard-delete)
      operationId: hardDeleteMunicipality
      tags: [municipalities]
      responses:
        '204':
          description: Município removido definitivamente (sem conteúdo)
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # PATCH /municipalities/:id/restore
  # ─────────────────────────────────────────────
  /municipalities/{id}/restore:
    parameters:
      - $ref: '#/components/parameters/MunicipalityID'

    patch:
      summary: Restaurar município removido da lixeira
      operationId: restoreMunicipality
      tags: [municipalities]
      responses:
        '200':
          description: Município restaurado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Passagem"
                  uf: "PB"
                  imageUrl: "https://example.com/brasoes/passagem-pb.png"
                  createdAt: "2025-01-15T10:30:00Z"
                  updatedAt: "2025-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # GET /municipalities/uf/:uf
  # ─────────────────────────────────────────────
  /municipalities/uf/{uf}:
    get:
      summary: Listar municípios por UF
      operationId: getMunicipalitiesByUF
      tags: [municipalities]
      parameters:
        - name: uf
          in: path
          required: true
          description: Sigla do estado (2 letras maiúsculas)
          schema:
            type: string
            minLength: 2
            maxLength: 2
            pattern: '^[A-Z]{2}$'
          example: PB
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Lista paginada filtrada por UF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
              example:
                success: true
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Passagem"
                    uf: "PB"
                    createdAt: "2025-01-15T10:30:00Z"
                    updatedAt: "2025-01-15T10:30:00Z"
                pagination:
                  page: 1
                  pageSize: 20
                  total: 1
                  totalPages: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # GET /health
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Parameters ─────────────────────────────
  parameters:
    MunicipalityID:
      name: id
      in: path
      required: true
      description: UUID v4 do município
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Page:
      name: page
      in: query
      required: false
      description: Número da página (começa em 1)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Itens por página (máx. 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20

  # ─── Schemas ────────────────────────────────
  schemas:
...
    Municipality:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          minLength: 2
          maxLength: 255
          example: "Passagem"
        uf:
          type: string
          minLength: 2
          maxLength: 2
          pattern: '^[A-Z]{2}$'
          example: "PB"
        imageUrl:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/brasoes/passagem-pb.png"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-01-15T10:30:00Z"
      required: [id, name, uf, createdAt, updatedAt]
    CreateMunicipalityInput:
      type: object
      required: [name, uf]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          example: "Passagem"
        uf:
          type: string
          minLength: 2
          maxLength: 2
          pattern: '^[A-Z]{2}$'
          example: "PB"
        imageUrl:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/brasoes/passagem-pb.png"
    UpdateMunicipalityInput:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          example: "Passagem"
        uf:
          type: string
          minLength: 2
          maxLength: 2
          pattern: '^[A-Z]{2}$'
          example: "PB"
        imageUrl:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/brasoes/passagem-pb.png"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Municipality'
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Municipality'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 42
        totalPages:
          type: integer
          example: 3
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "descrição do erro"
...
  responses:
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_uuid:
              summary: UUID malformado
              value:
                success: false
                error: "ID inválido"
            invalid_uf:
              summary: UF inexistente
              value:
                success: false
                error: "UF inválida"
            validation:
              summary: Falha de validação
              value:
                success: false
                error: "Key: 'CreateMunicipalityInput.Name' Error:Field validation for 'Name' failed on the 'required' tag"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "município não encontrado"
    Conflict:
      description: Conflito — nome já cadastrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "já existe um município com este nome"
    InternalError:
      description: Erro interno no servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "erro interno no servidor"